bot/requirements.txt
# Agent Pilot Bot - Dependencies
# Python 3.11+

# Telegram
python-telegram-bot==21.0.1
telethon==1.34.0

# Database
supabase==2.3.4

# HTTP & Async
aiohttp==3.9.3
httpx==0.26.0

# AI Providers
openai==1.12.0
anthropic==0.18.1

# Payments
stripe==8.4.0

# Utils
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.2.1
tenacity==8.2.3
cryptography==42.0.2

# Social Media
tweepy==4.14.0
instagrapi==2.0.0
python-wordpress-xmlrpc==2.3

# Media processing
Pillow==10.2.0
cloudinary==1.38.0

# Logging & Monitoring
structlog==24.1.0
sentry-sdk==1.40.0
